backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework-simplejwt==5.3.1
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
cloudinary==1.42.2
django-cloudinary-storage==0.3.0
